Move contract date formatting to tableMap
